move TFDatasetGenerator
