+++ rexify/features/extractor.py
@@ -1,13 +1,58 @@
 from typing import Any, Callable
 
+import numpy as np
 import pandas as pd
+import tensorflow as tf
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.pipeline import make_pipeline
 from sklearn.preprocessing import OrdinalEncoder
 
-from rexify.features.base import HasSchemaMixin, Serializable, TFDatasetGenerator
+from rexify.features.base import HasSchemaMixin, Serializable
 from rexify.features.transform import EventEncoder, Sequencer, TargetTransformer
 from rexify.schema import Schema
+from rexify.utils import get_target_id
+
+
+class _TFDatasetGenerator(HasSchemaMixin):
+    def make_dataset(self, X: pd.DataFrame) -> tf.data.Dataset:
+        ds = self._get_dataset(X)
+        ds = ds.map(self._get_header_fn())
+        return ds
+
+    def _get_dataset(self, data: pd.DataFrame) -> tf.data.Dataset:
+        return tf.data.Dataset.zip(
+            (
+                self._get_target_vector_dataset(data, self._schema, "user"),
+                self._get_target_vector_dataset(data, self._schema, "item"),
+                tf.data.Dataset.from_tensor_slices(
+                    np.stack(data["history"].values).astype(np.int32)
+                ),
+                tf.data.Dataset.from_tensor_slices(
+                    np.stack(data[self._schema.event_type].values).astype(np.float32)
+                ),
+            )
+        )
+
+    @staticmethod
+    def _get_target_vector_dataset(
+        data, schema: Schema, target: str
+    ) -> tf.data.Dataset:
+        return tf.data.Dataset.from_tensor_slices(
+            data.loc[:, get_target_id(schema, target)]
+            .values.reshape(-1)
+            .astype(np.int32)
+        )
+
+    @staticmethod
+    def _get_header_fn():
+        def header_fn(user_id, item_id, history, event):
+            return {
+                "query": {"user_id": user_id, "history": history},
+                "candidate": {"item_id": item_id},
+                "event": event,
+            }
+
+        return header_fn
 
 
 class _EventGenerator(BaseEstimator, TransformerMixin, HasSchemaMixin):
@@ -87,9 +132,13 @@ class _EventGenerator(BaseEstimator, TransformerMixin, HasSchemaMixin):
     def window_size(self):
         return self._window_size
 
+    @property
+    def ranking_features(self):
+        return self._ppl.steps[0][1].ranking_features
+
 
 class FeatureExtractor(
-    BaseEstimator, TransformerMixin, Serializable, TFDatasetGenerator
+    BaseEstimator, TransformerMixin, Serializable, _TFDatasetGenerator
 ):
 
     _event_gen: _EventGenerator
@@ -150,3 +199,7 @@ class FeatureExtractor(
     @property
     def model_params(self):
         return self._model_params
+
+    @property
+    def ranking_features(self):
+        return self._event_gen.ranking_features
